extract interface for board repository
